requirements.txt
discord.py>=2.3
aiosqlite>=0.19
python-dotenv>=1.0
tzdata>=2024.1
dateparser>=1.3.0
